requirements.txt
selenium
webdriver-manager
pytest
pyhamcrest
openpyxl
allure-pytest